% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acceleration_percentiles_cpp}
\alias{acceleration_percentiles_cpp}
\title{Calculate Acceleration Percentiles}
\usage{
acceleration_percentiles_cpp(data, probs)
}
\arguments{
\item{data}{NumericVector of acceleration values}

\item{probs}{NumericVector of probabilities (0-1)}
}
\value{
NumericVector of percentile values
}
\description{
Fast percentile calculation for acceleration distribution.
}
\keyword{internal}
