% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{activity_profile_cpp}
\alias{activity_profile_cpp}
\title{Calculate Complete Activity Profile (C++ Implementation)}
\usage{
activity_profile_cpp(
  data,
  epoch_length = 5,
  bin_size = 25,
  X_values = as.integer(c(2, 5, 10, 15, 30, 60, 120))
)
}
\arguments{
\item{data}{NumericVector of acceleration values in mg}

\item{epoch_length}{Epoch length in seconds}

\item{bin_size}{Bin size for intensity gradient}

\item{X_values}{X values for MX metrics}
}
\value{
List with all activity profile metrics
}
\description{
Combines intensity gradient and MX metrics in a single optimized function.
}
\keyword{internal}
