% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_agd.R
\name{analyze.agd.file}
\alias{analyze.agd.file}
\title{Analyze AGD File}
\usage{
analyze.agd.file(...)
}
\arguments{
\item{...}{Arguments passed to \code{canhrActi()}}
}
\description{
Alias for \code{\link{canhrActi}}.
}
