% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{backend_info}
\alias{backend_info}
\title{Get Backend Information}
\usage{
backend_info()
}
\value{
Invisible list with backend info
}
\description{
Display information about computational backends.
}
