% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{butte_preschool}
\alias{butte_preschool}
\title{Apply Butte (2014) Preschooler Cut Points}
\usage{
butte_preschool(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of vertical-axis CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for preschoolers (mean age 4.5) validated against room calorimetry,
for the ActiGraph vertical axis.
}
\details{
Butte (2014) cutpoints, ActiGraph x-axis:
\itemize{
\item Sedentary: 0-239 CPM
\item Light: 240-2119 CPM
\item Moderate: 2120-4449 CPM
\item Vigorous: >= 4450 CPM
}

The paper gives one set for preschoolers, with no cut by age; for school-age
children see \code{\link{evenson}}.
}
\references{
Butte NF, Wong WW, Lee JS, Adolph AL, Puyau MR, Zakeri IF. (2014). Prediction
of energy expenditure and physical activity in preschoolers. Med Sci Sports
Exerc, 46(7), 1216-1226.
}
\seealso{
Other children cut-points:
\code{\link[=evenson]{evenson()}},
\code{\link[=mattocks]{mattocks()}},
\code{\link[=pate_preschool]{pate_preschool()}},
\code{\link[=puyau]{puyau()}},
\code{\link[=romanzini]{romanzini()}}
}
\concept{children cut-points}
