% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_expenditure.R
\name{calculate.average.mets}
\alias{calculate.average.mets}
\title{Calculate Average METs}
\usage{
calculate.average.mets(mets, wear_time, timestamp)
}
\arguments{
\item{mets}{Numeric vector of METs values}

\item{wear_time}{Logical vector indicating wear time}

\item{timestamp}{POSIXct vector of timestamps}
}
\value{
List with average_mets, daily_average_mets
}
\description{
Calculate Average METs
}
