% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{calculate_alpha}
\alias{calculate_alpha}
\title{Calculate Power-Law Alpha (Fragmentation Index)}
\usage{
calculate_alpha(durations, xmin = 1)
}
\arguments{
\item{durations}{Numeric vector of bout durations}

\item{xmin}{Minimum value for power-law fitting (default 1)}
}
\value{
Estimated alpha value
}
\description{
Estimates the power-law exponent for bout duration distribution using
maximum likelihood estimation.
}
\details{
Uses the Hill estimator (MLE) for power-law exponent:
alpha = 1 + n / sum(log(x / xmin))
}
\keyword{internal}
