% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{calculate_gini}
\alias{calculate_gini}
\title{Calculate Gini Coefficient}
\usage{
calculate_gini(x)
}
\arguments{
\item{x}{Numeric vector of values}
}
\value{
Gini coefficient (0 = perfect equality, 1 = perfect inequality)
}
\description{
Computes the Gini coefficient for bout duration inequality.
}
\keyword{internal}
