% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_MX_cpp}
\alias{calculate_MX_cpp}
\title{Calculate MX Metrics (C++ Implementation)}
\usage{
calculate_MX_cpp(data, X_values, epoch_length = 5)
}
\arguments{
\item{data}{NumericVector of acceleration values in mg}

\item{X_values}{IntegerVector of time values in minutes (e.g., c(2, 5, 10, 30, 60))}

\item{epoch_length}{Epoch length in seconds (default: 5)}
}
\value{
Named NumericVector of MX values
}
\description{
High-performance calculation of MX metrics (acceleration exceeded during
most active X minutes).
}
\keyword{internal}
