% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{CANHR.Cutpoints}
\alias{CANHR.Cutpoints}
\title{Apply CANHR Custom Cut Points}
\usage{
CANHR.Cutpoints(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of counts per minute}
}
\value{
Ordered factor with intensity levels
}
\description{
Custom cut-points developed by the Center for Alaska Native Health Research.
}
