% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canhrActi-package.R
\docType{package}
\name{canhrActi-package}
\alias{canhrActi-package}
\title{canhrActi: Activity, Sleep and Circadian Analysis of Accelerometer Data}
\description{
Accelerometer analysis for physical activity, sleep, sedentary behavior and circadian rhythm research, from the Center for Alaska Native Health Research (CANHR). Reads 'ActiGraph' count files (.agd) and raw recordings ('ActiGraph' .gt3x, 'Axivity' .cwa, 'GENEActiv' .bin). Count results are compatible with 'ActiLife', the raw pipeline gives the same results as 'GGIR', and an interactive dashboard runs the full analysis without writing code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdazadda/canhrActi}
  \item Report bugs at \url{https://github.com/rdazadda/canhrActi/issues}
}

}
\author{
\strong{Maintainer}: Raymond Dacosta Azadda \email{rdazadda@alaska.edu} (\href{https://orcid.org/0009-0002-4384-4556}{ORCID})

Authors:
\itemize{
  \item Raymond Dacosta Azadda \email{rdazadda@alaska.edu} (\href{https://orcid.org/0009-0002-4384-4556}{ORCID})
  \item AK CEAL Team
  \item Stacy Rasmus
}

}
\keyword{internal}
