% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{chandler}
\alias{chandler}
\title{Apply Chandler (2016) Preschooler Cut Points}
\usage{
chandler(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for preschoolers validated with indirect calorimetry.
Specifically developed for 3-5 year olds.
}
\details{
Chandler (2016) cutpoints for preschoolers (15-sec epochs):
\itemize{
\item Sedentary: 0-5 counts/15s (0-20 CPM)
\item Light: 6-404 counts/15s (21-1616 CPM)
\item Moderate: 405-810 counts/15s (1617-3240 CPM)
\item Vigorous: >= 811 counts/15s (>= 3241 CPM)
}

Note: Values shown are converted to CPM from original 15-second epoch values.
}
\references{
Chandler JL, et al. (2016). Classification of physical activity intensities
using a wrist-worn accelerometer in 8-12-year-old children.
Pediatr Obes, 11(2), 120-127.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{copeland}()},
\code{\link{evenson}()},
\code{\link{mattocks}()},
\code{\link{pate_preschool}()},
\code{\link{puyau}()},
\code{\link{romanzini}()}
}
\concept{children cut-points}
