% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mets_algorithms.R
\name{classify.mets}
\alias{classify.mets}
\title{Classify METs into Intensity Categories}
\usage{
classify.mets(mets)
}
\arguments{
\item{mets}{Numeric vector of METs values}
}
\value{
Factor vector with intensity categories (sedentary, light, moderate, vigorous, very_vigorous)
}
\description{
Converts METs values into standard physical activity intensity levels.
}
