% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{confusion.metrics}
\alias{confusion.metrics}
\title{Confusion Matrix Metrics (binary)}
\usage{
confusion.metrics(reference, predicted)
}
\arguments{
\item{reference}{Logical or numeric vector of true values}

\item{predicted}{Logical or numeric vector of predicted values}
}
\value{
List with sensitivity, specificity, accuracy, precision, F1
}
\description{
Confusion Matrix Metrics (binary)
}
