% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{copeland}
\alias{copeland}
\title{Apply Copeland (2009) Children Cut Points}
\usage{
copeland(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for children (13-14 years) validated in free-living conditions.
}
\details{
Copeland (2009) cutpoints:
\itemize{
\item Sedentary: 0-100 CPM
\item Light: 101-2220 CPM
\item Moderate: 2221-6130 CPM
\item Vigorous: >= 6131 CPM
}
}
\references{
Copeland JL, Esliger DW. (2009). Accelerometer assessment of physical
activity in active, healthy older adults. J Aging Phys Act, 17(1), 17-30.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{chandler}()},
\code{\link{evenson}()},
\code{\link{mattocks}()},
\code{\link{pate_preschool}()},
\code{\link{puyau}()},
\code{\link{romanzini}()}
}
\concept{children cut-points}
