% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{.calculate.alpha.simple}
\alias{.calculate.alpha.simple}
\title{Simple Power-Law Alpha (Hill Estimator)}
\usage{
.calculate.alpha.simple(durations, xmin = 1)
}
\arguments{
\item{durations}{Numeric vector of bout durations}

\item{xmin}{Minimum value for fitting (default 1)}
}
\value{
Estimated alpha value
}
\description{
Simple Power-Law Alpha (Hill Estimator)
}
\keyword{internal}
