% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{.calculate.daily.fragmentation}
\alias{.calculate.daily.fragmentation}
\title{Calculate Daily Fragmentation Patterns}
\usage{
.calculate.daily.fragmentation(intensity, timestamps, wear_time, epoch_length)
}
\description{
Calculate Daily Fragmentation Patterns
}
\keyword{internal}
