% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sleep.R
\name{.canhrActi.sleep.single}
\alias{.canhrActi.sleep.single}
\title{Single File Sleep Analysis (Internal)}
\usage{
.canhrActi.sleep.single(
  agd_file_path,
  sleep_algorithm,
  apply_rescoring,
  detect_sleep_period,
  bedtime_start,
  wake_time_end,
  min_sleep_period,
  max_sleep_period,
  min_nonzero_epochs,
  verbose = TRUE
)
}
\description{
Single File Sleep Analysis (Internal)
}
\keyword{internal}
