% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sleep.R
\name{.create.actilife.batch.details}
\alias{.create.actilife.batch.details}
\title{Create ActiLife Batch Details (Internal)}
\usage{
.create.actilife.batch.details(results.list)
}
\arguments{
\item{results.list}{List of sleep results}
}
\value{
Data frame in ActiLife format
}
\description{
Creates the BatchSleepExportDetails.csv file with one row per sleep period.
}
\keyword{internal}
