% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{.create.empty.bouts.df}
\alias{.create.empty.bouts.df}
\title{Create Empty Bouts Data Frame}
\usage{
.create.empty.bouts.df()
}
\description{
Create Empty Bouts Data Frame
}
\keyword{internal}
