% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{.create.empty.fragmentation.result}
\alias{.create.empty.fragmentation.result}
\title{Create Empty Fragmentation Result}
\usage{
.create.empty.fragmentation.result(
  total_sedentary_min,
  total_wear_min,
  timestamps
)
}
\description{
Create Empty Fragmentation Result
}
\keyword{internal}
