% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_detection.R
\name{.detect.bouts.drop.time}
\alias{.detect.bouts.drop.time}
\title{Detect Bouts Using Drop Time Allowance Method}
\usage{
.detect.bouts.drop.time(is_mvpa, min_bout_length, drop_time_allowance)
}
\arguments{
\item{is_mvpa}{Logical vector indicating MVPA epochs}

\item{min_bout_length}{Minimum bout length in minutes}

\item{drop_time_allowance}{Maximum consecutive minutes below MVPA allowed}
}
\value{
List of bouts, each with start and end indices
}
\description{
Internal function implementing NHANES/CDC standard bout detection.
}
\keyword{internal}
