% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sleep.R
\name{.format.actilife.datetime}
\alias{.format.actilife.datetime}
\title{Format DateTime for ActiLife (Internal)}
\usage{
.format.actilife.datetime(dt)
}
\arguments{
\item{dt}{POSIXct datetime}
}
\value{
Character string in ActiLife format
}
\description{
Converts R datetime to ActiLife format: M/D/YYYY H:MM:SS AM/PM
}
\keyword{internal}
