% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sleep.R
\name{.format.average.time}
\alias{.format.average.time}
\title{Format Average Time for ActiLife (Internal)}
\usage{
.format.average.time(times)
}
\arguments{
\item{times}{Vector of POSIXct datetimes}
}
\value{
Character string in ActiLife time format
}
\description{
Converts average of POSIXct times to ActiLife format: H:MM AM/PM
Uses circular mean to handle times around midnight properly
}
\keyword{internal}
