% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sleep.R
\name{.format.gender}
\alias{.format.gender}
\title{Format Gender for ActiLife (Internal Helper)}
\usage{
.format.gender(sex)
}
\arguments{
\item{sex}{Sex character code}
}
\value{
Formatted gender string
}
\description{
Converts sex code to ActiLife format.
}
\keyword{internal}
