% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{.is.agd.file}
\alias{.is.agd.file}
\title{Check if File is AGD Format}
\usage{
.is.agd.file(filepath)
}
\arguments{
\item{filepath}{Path to file}
}
\value{
Logical. TRUE if file has .agd extension
}
\description{
Check if File is AGD Format
}
\keyword{internal}
