% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x.R
\name{.is.gt3x.file}
\alias{.is.gt3x.file}
\title{Check if File is GT3X Format}
\usage{
.is.gt3x.file(filepath)
}
\arguments{
\item{filepath}{Character. Path to file}
}
\value{
Logical. TRUE if file has .gt3x extension
}
\description{
Quick check to determine if a file is in .gt3x format based on extension.
}
\keyword{internal}
