% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{.read.agd.raw}
\alias{.read.agd.raw}
\title{Read Raw AGD File (Internal)}
\usage{
.read.agd.raw(con, settings, epoch_length = 60, verbose = TRUE)
}
\arguments{
\item{con}{Database connection}

\item{settings}{Settings data frame}

\item{epoch_length}{Epoch length in seconds}

\item{verbose}{Print progress messages?}
}
\value{
List with data and settings
}
\description{
Reads raw acceleration data from AGD file and converts to activity counts.
}
\keyword{internal}
