% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_devices.R
\name{.read.agd.unified}
\alias{.read.agd.unified}
\title{Read ActiGraph AGD File (Unified)}
\usage{
.read.agd.unified(filepath, verbose = TRUE)
}
\arguments{
\item{filepath}{Path to .agd file}

\item{verbose}{Print messages?}
}
\value{
Standardized data list
}
\description{
Read ActiGraph AGD File (Unified)
}
\keyword{internal}
