% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_devices.R
\name{.read.axivity.ggirread}
\alias{.read.axivity.ggirread}
\title{Read Axivity using GGIRread}
\usage{
.read.axivity.ggirread(filepath, epoch_length, verbose)
}
\description{
Read Axivity using GGIRread
}
\keyword{internal}
