% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_devices.R
\name{.read.axivity.readcwa}
\alias{.read.axivity.readcwa}
\title{Read Axivity using read.cwa}
\usage{
.read.axivity.readcwa(filepath, epoch_length, verbose)
}
\description{
Read Axivity using read.cwa
}
\keyword{internal}
