% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_devices.R
\name{.read.gt3x.unified}
\alias{.read.gt3x.unified}
\title{Read ActiGraph GT3X File (Unified)}
\usage{
.read.gt3x.unified(filepath, epoch_length = 60, verbose = TRUE, ...)
}
\arguments{
\item{filepath}{Path to .gt3x file}

\item{epoch_length}{Epoch length in seconds}

\item{verbose}{Print messages?}

\item{...}{Additional arguments}
}
\value{
Standardized data list
}
\description{
Read ActiGraph GT3X File (Unified)
}
\keyword{internal}
