% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{example_agd}
\alias{example_agd}
\title{Get Path to Example AGD Files}
\usage{
example_agd(file = 1)
}
\arguments{
\item{file}{Index, part of a file name, or "list" to see available files.}
}
\value{
Character. Full path to the example AGD file.
}
\description{
The example recordings are the dashboard's sample files.
}
\examples{
example_agd()
example_agd("list")
agd_path <- example_agd(1)
}
