% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_canhrActi.R
\name{export_daily_detailed_internal}
\alias{export_daily_detailed_internal}
\title{Internal Daily Detailed Export (Returns Data Frame)}
\usage{
export_daily_detailed_internal(analysis_results)
}
\description{
Internal Daily Detailed Export (Returns Data Frame)
}
\keyword{internal}
