% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_canhrActi.R
\name{export_hourly_detailed_internal}
\alias{export_hourly_detailed_internal}
\title{Internal Hourly Detailed Export (Returns Data Frame)}
\usage{
export_hourly_detailed_internal(analysis_results)
}
\description{
Internal Hourly Detailed Export (Returns Data Frame)
}
\keyword{internal}
