% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_canhrActi.R
\name{export_summary_internal}
\alias{export_summary_internal}
\title{Internal Summary Export (Returns Data Frame)}
\usage{
export_summary_internal(analysis_results)
}
\description{
Internal Summary Export (Returns Data Frame)
}
\keyword{internal}
