% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{extract.subject.info}
\alias{extract.subject.info}
\title{Extract Subject Information from AGD Settings}
\usage{
extract.subject.info(agd_data)
}
\arguments{
\item{agd_data}{List returned from read.agd()}
}
\value{
List with subject information
}
\description{
Extract Subject Information from AGD Settings
}
\keyword{internal}
