% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter.time.range}
\alias{filter.time.range}
\title{Filter Accelerometer Data by Time Range}
\usage{
filter.time.range(accel.data, start.time, end.time)
}
\arguments{
\item{accel.data}{Data frame with timestamp column}

\item{start.time}{POSIXct or character. Start time for filter}

\item{end.time}{POSIXct or character. End time for filter}
}
\value{
Data frame filtered to the specified time range
}
\description{
Filter Accelerometer Data by Time Range
}
