% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intensity_gradient_cpp}
\alias{intensity_gradient_cpp}
\title{Calculate Intensity Gradient (C++ Implementation)}
\usage{
intensity_gradient_cpp(data, epoch_length = 5, bin_size = 25, max_value = 4000)
}
\arguments{
\item{data}{NumericVector of acceleration values in mg}

\item{epoch_length}{Epoch length in seconds (default: 5)}

\item{bin_size}{Bin width in mg (default: 25)}

\item{max_value}{Maximum value for binning (default: 4000)}
}
\value{
List with gradient, intercept, r_squared, average_acceleration
}
\description{
High-performance calculation of the intensity gradient using log-log
linear regression on acceleration bin data.
}
\keyword{internal}
