% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{IS_cpp}
\alias{IS_cpp}
\title{Calculate Interdaily Stability (C++ Implementation)}
\usage{
IS_cpp(hourly_data, hours_per_day = 24L)
}
\arguments{
\item{hourly_data}{Numeric vector of hourly activity data}

\item{hours_per_day}{Hours per day (default: 24)}
}
\value{
Numeric IS value (0-1, higher = more stable)
}
\description{
Measures day-to-day consistency of activity patterns (0-1).
}
