% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{list_cutpoints}
\alias{list_cutpoints}
\title{List Available Cut-Point Algorithms}
\usage{
list_cutpoints(category = "all")
}
\arguments{
\item{category}{Filter by category: "all", "adult", "children",
"older_adult", "triaxial"}
}
\value{
Data frame with algorithm information
}
\description{
Returns information about all available cut-point algorithms
including their target population and reference.
}
\examples{
# List all algorithms
list_cutpoints()

# List only children algorithms
list_cutpoints("children")

}
