% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{mvpa_bouts_cpp}
\alias{mvpa_bouts_cpp}
\title{Detect MVPA Bouts}
\usage{
mvpa_bouts_cpp(
  counts,
  moderate_threshold = 1952L,
  min_bout_length = 10L,
  drop_time = 2L,
  use_80_percent_rule = FALSE
)
}
\arguments{
\item{counts}{Activity counts}

\item{moderate_threshold}{MVPA threshold (default: 1952)}

\item{min_bout_length}{Min bout epochs (default: 10)}

\item{drop_time}{Allowed drop epochs (default: 2)}

\item{use_80_percent_rule}{Use 80\% rule? (default: FALSE)}
}
\value{
List with bout details and statistics
}
\description{
Detect MVPA Bouts
}
