% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity.R
\name{plot_daily_summary}
\alias{plot_daily_summary}
\title{Plot Daily Summary}
\usage{
plot_daily_summary(x)
}
\arguments{
\item{x}{A \code{canhrActi_analysis} object from \code{canhrActi()}}
}
\value{
A ggplot2 object
}
\description{
Creates a multi-panel visualization showing daily patterns of activity metrics.
}
\examples{
\dontrun{
results <- canhrActi("participant.agd")
plot_daily_summary(results)
}

}
