% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{plot_hourly_heatmap}
\alias{plot_hourly_heatmap}
\title{Plot Hourly Fragmentation Heatmap}
\usage{
plot_hourly_heatmap(frag, metric = "bout_count")
}
\arguments{
\item{frag}{A canhrActi_fragmentation object}

\item{metric}{Which metric to display: "bout_count", "duration", or "fragmentation"}
}
\value{
A ggplot2 object
}
\description{
Creates a heatmap visualization of sedentary fragmentation by hour of day
and day of recording.
}
