% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity.R
\name{plot_intensity_distribution}
\alias{plot_intensity_distribution}
\title{Plot Intensity Distribution}
\usage{
plot_intensity_distribution(x, style = c("bar", "pie"))
}
\arguments{
\item{x}{A \code{canhrActi_analysis} object from \code{canhrActi()}}

\item{style}{Character. Either "bar" (default) or "pie"}
}
\value{
A ggplot2 object
}
\description{
Creates a visual summary of time spent in each intensity level.
}
