% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_agd.R
\name{print.canhrActi_analysis}
\alias{print.canhrActi_analysis}
\title{Print Method for canhrActi Analysis Results}
\usage{
\method{print}{canhrActi_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{canhrActi_analysis}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for canhrActi Analysis Results
}
