% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_analysis.R
\name{print.canhrActi_batch}
\alias{print.canhrActi_batch}
\title{Print Method for Batch Results}
\usage{
\method{print}{canhrActi_batch}(x, ...)
}
\arguments{
\item{x}{An object of class canhrActi_batch}

\item{...}{Additional arguments}
}
\description{
Print Method for Batch Results
}
