% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{print.canhrActi_calibration}
\alias{print.canhrActi_calibration}
\title{Print Method for Calibration Results}
\usage{
\method{print}{canhrActi_calibration}(x, ...)
}
\arguments{
\item{x}{Object of class 'canhrActi_calibration'}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for Calibration Results
}
