% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{print.canhrActi_circadian}
\alias{print.canhrActi_circadian}
\title{Print Method for Circadian Rhythm Results}
\usage{
\method{print}{canhrActi_circadian}(x, ...)
}
\arguments{
\item{x}{Object of class 'canhrActi_circadian'}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for Circadian Rhythm Results
}
