% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{print.canhrActi_cosinor}
\alias{print.canhrActi_cosinor}
\title{Print method for cosinor analysis}
\usage{
\method{print}{canhrActi_cosinor}(x, ...)
}
\arguments{
\item{x}{canhrActi_cosinor object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cosinor analysis
}
