% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{print.canhrActi_cosinor_extended}
\alias{print.canhrActi_cosinor_extended}
\title{Print method for extended cosinor analysis}
\usage{
\method{print}{canhrActi_cosinor_extended}(x, ...)
}
\arguments{
\item{x}{canhrActi_cosinor_extended object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for extended cosinor analysis
}
