% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{print.canhrActi_fragmentation}
\alias{print.canhrActi_fragmentation}
\title{Print Method for Fragmentation Results}
\usage{
\method{print}{canhrActi_fragmentation}(x, ...)
}
\arguments{
\item{x}{A canhrActi_fragmentation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Fragmentation Results
}
