% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_canhrActi.R
\name{print.canhrActi_ibi_analysis}
\alias{print.canhrActi_ibi_analysis}
\title{Print Inter-Bout Interval Analysis}
\usage{
\method{print}{canhrActi_ibi_analysis}(x, ...)
}
\arguments{
\item{x}{Object from analyze_inter_bout_intervals()}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Inter-Bout Interval Analysis
}
