% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_fractal.R
\name{print.canhrActi_mse}
\alias{print.canhrActi_mse}
\title{Print Method for Multiscale Entropy Results}
\usage{
\method{print}{canhrActi_mse}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"canhrActi_mse"}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Multiscale Entropy Results
}
