% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_pipeline.R
\name{print.canhrActi_raw}
\alias{print.canhrActi_raw}
\title{Print Method for a Raw Accelerometer Recording}
\usage{
\method{print}{canhrActi_raw}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw object.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Raw Accelerometer Recording
}
