% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_calibrate.R
\name{print.canhrActi_raw_calibration}
\alias{print.canhrActi_raw_calibration}
\title{Print method for a raw calibration}
\usage{
\method{print}{canhrActi_raw_calibration}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_calibration.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print method for a raw calibration
}
