% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_impute.R
\name{print.canhrActi_raw_imputed}
\alias{print.canhrActi_raw_imputed}
\title{Print method for an imputed short-epoch object}
\usage{
\method{print}{canhrActi_raw_imputed}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_imputed.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print method for an imputed short-epoch object
}
