% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_inspect.R
\name{print.canhrActi_raw_info}
\alias{print.canhrActi_raw_info}
\title{Print a Raw File Inspection}
\usage{
\method{print}{canhrActi_raw_info}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_info object from raw.inspect().}

\item{...}{Ignored.}
}
\value{
x, invisibly.
}
\description{
Print a Raw File Inspection
}
