% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_getmeta.R
\name{print.canhrActi_raw_meta}
\alias{print.canhrActi_raw_meta}
\title{Print Method for a Raw Epoch Table Object}
\usage{
\method{print}{canhrActi_raw_meta}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_meta.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Raw Epoch Table Object
}
