% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_sleep_nights.R
\name{print.canhrActi_raw_nights}
\alias{print.canhrActi_raw_nights}
\title{Print Method for a Night Table}
\usage{
\method{print}{canhrActi_raw_nights}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_nights.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Night Table
}
