% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_sleep_detect.R
\name{print.canhrActi_raw_sib}
\alias{print.canhrActi_raw_sib}
\title{Print Method for a Sustained Inactivity Bout Object}
\usage{
\method{print}{canhrActi_raw_sib}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_sib.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Sustained Inactivity Bout Object
}
