% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_sleep_part3.R
\name{print.canhrActi_raw_sleep}
\alias{print.canhrActi_raw_sleep}
\title{Print Method for a Part-3 Sleep Object}
\usage{
\method{print}{canhrActi_raw_sleep}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_sleep.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Part-3 Sleep Object
}
