% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_sleep_report.R
\name{print.canhrActi_raw_sleep_report}
\alias{print.canhrActi_raw_sleep_report}
\title{Print Method for a Sleep Report}
\usage{
\method{print}{canhrActi_raw_sleep_report}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_sleep_report.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Sleep Report
}
