% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_timeuse.R
\name{print.canhrActi_raw_timeuse}
\alias{print.canhrActi_raw_timeuse}
\title{Print Method for a Time-Use Object}
\usage{
\method{print}{canhrActi_raw_timeuse}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_timeuse.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print Method for a Time-Use Object
}
