% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_timeuse_report.R
\name{print.canhrActi_raw_timeuse_report}
\alias{print.canhrActi_raw_timeuse_report}
\title{Print a Part-5 Report}
\usage{
\method{print}{canhrActi_raw_timeuse_report}(x, ...)
}
\arguments{
\item{x}{A \code{canhrActi_raw_timeuse_report}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Part-5 Report
}
