% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_weardec.R
\name{print.canhrActi_raw_wear}
\alias{print.canhrActi_raw_wear}
\title{Print method for a wear decision object}
\usage{
\method{print}{canhrActi_raw_wear}(x, ...)
}
\arguments{
\item{x}{A canhrActi_raw_wear.}

\item{...}{Not used.}
}
\value{
x, invisibly.
}
\description{
Print method for a wear decision object
}
