% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sleep.R
\name{print.canhrActi_sleep}
\alias{print.canhrActi_sleep}
\title{Print Method for Single Sleep Analysis}
\usage{
\method{print}{canhrActi_sleep}(x, ...)
}
\arguments{
\item{x}{An object of class canhrActi_sleep}

\item{...}{Additional arguments}
}
\description{
Print Method for Single Sleep Analysis
}
