% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sleep.R
\name{print.canhrActi_sleep_batch}
\alias{print.canhrActi_sleep_batch}
\title{Print Method for Sleep Batch Results}
\usage{
\method{print}{canhrActi_sleep_batch}(x, ...)
}
\arguments{
\item{x}{An object of class canhrActi_sleep_batch}

\item{...}{Additional arguments}
}
\description{
Print Method for Sleep Batch Results
}
