% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.canhrActi_valid_days}
\alias{print.canhrActi_valid_days}
\title{Print Method for Valid Days Results}
\usage{
\method{print}{canhrActi_valid_days}(x, ...)
}
\description{
Print Method for Valid Days Results
}
