% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{print.canhrActi_validation}
\alias{print.canhrActi_validation}
\title{Print Method for Validation Metrics}
\usage{
\method{print}{canhrActi_validation}(x, ...)
}
\description{
Print Method for Validation Metrics
}
