% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_advanced.R
\name{quick_plots}
\alias{quick_plots}
\title{Quick Visualization Function}
\usage{
quick_plots(result, plots = c("timeline", "heatmap", "intensity"))
}
\arguments{
\item{result}{A canhrActi analysis result object}

\item{plots}{Character vector of plot types to generate}
}
\value{
A list of ggplot2 objects
}
\description{
Generates multiple standard visualizations from a canhrActi analysis result.
}
