% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{rolling_sd}
\alias{rolling_sd}
\title{Rolling Standard Deviation (C++ Implementation)}
\usage{
rolling_sd(x, window)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Window size}
}
\value{
Numeric vector of rolling SDs
}
\description{
Rolling SD using Welford's algorithm.
}
