% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample.rate}
\alias{sample.rate}
\title{Detect Sampling Frequency from timestamps}
\usage{
sample.rate(timestamps)
}
\arguments{
\item{timestamps}{POSIXct vector of timestamps}
}
\value{
Numeric. Detected sampling frequency in Hz
}
\description{
Detect Sampling Frequency from timestamps
}
