% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{sasaki_vm3}
\alias{sasaki_vm3}
\title{Apply Sasaki VM3 (2011) Cut Points}
\usage{
sasaki_vm3(vm_counts_per_minute)
}
\arguments{
\item{vm_counts_per_minute}{Numeric vector of vector magnitude CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Triaxial cut-points using Vector Magnitude (VM) counts from all three axes.
More accurate than uniaxial for capturing multi-planar movement.
}
\details{
Vector Magnitude is calculated as: VM = sqrt(x^2 + y^2 + z^2)

Sasaki (2011) VM cutpoints:
\itemize{
\item Sedentary: 0-199 VM CPM
\item Light: 200-2689 VM CPM
\item Moderate: 2690-6166 VM CPM
\item Vigorous: 6167-9642 VM CPM
\item Very Vigorous: >= 9643 VM CPM
}
}
\references{
Sasaki JE, et al. (2011). Validation and comparison of ActiGraph activity
monitors. J Sci Med Sport, 14(5), 411-416.
}
\seealso{
Other triaxial cut-points:
\code{\link[=freedson_vm3]{freedson_vm3()}},
\code{\link[=santos_lozano]{santos_lozano()}}
}
\concept{triaxial cut-points}
