% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_canhrActi}
\alias{scale_color_canhrActi}
\title{ggplot2 Scale for canhrActi Colors}
\usage{
scale_color_canhrActi(type = "categorical", ...)
}
\arguments{
\item{type}{Character. Palette type (see canhrActi_colors)}

\item{...}{Additional arguments passed to ggplot2::scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
Discrete color scale using the canhrActi palette.
}
