% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_canhrActi}
\alias{scale_fill_canhrActi}
\title{ggplot2 Fill Scale for canhrActi Colors}
\usage{
scale_fill_canhrActi(type = "categorical", ...)
}
\arguments{
\item{type}{Character. Palette type (see canhrActi_colors)}

\item{...}{Additional arguments passed to ggplot2::scale_fill_manual}
}
\value{
A ggplot2 scale object
}
\description{
Discrete fill scale using the canhrActi palette.
}
