% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{sedentary_bouts_cpp}
\alias{sedentary_bouts_cpp}
\title{Detect Sedentary Bouts}
\usage{
sedentary_bouts_cpp(
  counts,
  sedentary_threshold = 100L,
  min_bout_length = 1L,
  wear = integer()
)
}
\arguments{
\item{counts}{Activity counts}

\item{sedentary_threshold}{Sedentary threshold (default: 100)}

\item{min_bout_length}{Min bout length (default: 1)}

\item{wear}{Wear time vector (optional)}
}
\value{
List with bout details and statistics
}
\description{
Detect Sedentary Bouts
}
