% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_canhrActi_theme}
\alias{set_canhrActi_theme}
\title{Set canhrActi Theme as Default}
\usage{
set_canhrActi_theme(...)
}
\arguments{
\item{...}{Arguments passed to theme_canhrActi()}
}
\description{
Sets theme_canhrActi() as the default ggplot2 theme for the session.
}
