% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{sri_cpp}
\alias{sri_cpp}
\title{Sleep Regularity Index (SRI)}
\usage{
sri_cpp(sleep, epochs_per_day = 1440L)
}
\arguments{
\item{sleep}{Sleep scores}

\item{epochs_per_day}{Epochs per day (default: 1440)}
}
\value{
SRI value (-100 to 100)
}
\description{
Sleep Regularity Index (SRI)
}
\references{
Phillips AJK, et al. (2017). Scientific Reports.
}
