% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_counting.R
\name{steps.auto}
\alias{steps.auto}
\title{Count Steps Using Autocorrelation Method}
\usage{
steps.auto(
  acceleration,
  sampling_freq = 60,
  min_cadence = 60,
  max_cadence = 200
)
}
\arguments{
\item{acceleration}{Numeric vector of vertical axis acceleration}

\item{sampling_freq}{Numeric. Sampling frequency in Hz}

\item{min_cadence, max_cadence}{Expected cadence bounds (spm)}
}
\value{
Integer. Estimated total number of steps
}
\description{
Count Steps Using Autocorrelation Method
}
