% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_expenditure.R
\name{summarize.energy.expenditure}
\alias{summarize.energy.expenditure}
\title{Summarize Energy Expenditure by Intensity}
\usage{
summarize.energy.expenditure(kcal_per_epoch, intensity, wear_time)
}
\arguments{
\item{kcal_per_epoch}{Numeric vector of kcal per epoch}

\item{intensity}{Factor vector of intensity classifications}

\item{wear_time}{Logical vector indicating wear time}
}
\value{
Data frame with intensity, total_kcal, percent_kcal
}
\description{
Summarize Energy Expenditure by Intensity
}
