% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_detection.R
\name{summarize.mvpa.bouts}
\alias{summarize.mvpa.bouts}
\title{Summarize MVPA Bouts}
\usage{
summarize.mvpa.bouts(bouts)
}
\arguments{
\item{bouts}{Data frame returned from \code{detect.mvpa.bouts}}
}
\value{
List with summary statistics:
\itemize{
\item \code{total_bouts} - Number of bouts detected
\item \code{total_bouted_mvpa} - Total MVPA minutes in bouts
\item \code{mean_bout_length} - Average bout duration
\item \code{median_bout_length} - Median bout duration
\item \code{min_bout_length} - Shortest bout
\item \code{max_bout_length} - Longest bout
}
}
\description{
Calculates summary statistics for detected MVPA bouts.
}
