% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_agd.R
\name{summary.canhrActi_analysis}
\alias{summary.canhrActi_analysis}
\title{Summary Method for canhrActi Analysis Results}
\usage{
\method{summary}{canhrActi_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class \code{canhrActi_analysis}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for canhrActi Analysis Results
}
