% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{synthetic}
\alias{synthetic}
\title{Generate Synthetic Accelerometer Data}
\usage{
synthetic(
  duration.hours = 24,
  sampling.freq = 60,
  add.activity = TRUE,
  n.bouts = 10,
  step.freq = 2
)
}
\description{
Generate Synthetic Accelerometer Data
}
