% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_timestamps}
\alias{validate_timestamps}
\title{Validate Timestamps}
\usage{
validate_timestamps(ts)
}
\arguments{
\item{ts}{Timestamps to validate}
}
\description{
Validate Timestamps
}
\keyword{internal}
