% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation.metrics}
\alias{validation.metrics}
\title{Calculate Validation Metrics}
\usage{
validation.metrics(reference, test)
}
\arguments{
\item{reference}{Numeric vector of reference/gold standard values}

\item{test}{Numeric vector of test/new method values}
}
\value{
List of bias, MAE, MAPE, RMSE, r, ICC, LoA, n
}
\description{
Calculate Validation Metrics
}
