% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{agd.counts}
\alias{agd.counts}
\title{Extract Counts from .agd Data}
\usage{
agd.counts(agd_data, convert.timestamps = TRUE)
}
\arguments{
\item{agd_data}{List returned from read.agd()}

\item{convert.timestamps}{Logical. Convert ActiGraph timestamps to POSIXct (default: TRUE)}
}
\value{
Data frame with counts per minute and timestamps
}
\description{
Extract Counts from .agd Data
}
