% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_analysis.R
\name{batch.config}
\alias{batch.config}
\title{Create Batch Processing Configuration}
\usage{
batch.config(
  wear = NULL,
  intensity = NULL,
  min_wear = NULL,
  axis = NULL,
  mets = NULL,
  mets_algo = NULL,
  sleep = NULL,
  age = NULL,
  fragmentation = NULL,
  circadian = NULL,
  exclude_sleep = NULL,
  parallel = NULL,
  cores = NULL
)
}
\arguments{
\item{wear}{Wear time algorithm: "choi", "troiano", "CANHR2025"}

\item{intensity}{Intensity algorithm: "freedson1998", "CANHR"}

\item{min_wear}{Minimum wear hours (default: 10)}

\item{axis}{Axis: "axis1", "vector_magnitude"}

\item{mets}{Calculate METs? (default: TRUE)}

\item{mets_algo}{METs algorithm}

\item{sleep}{Sleep algorithm: "cole_kripke", "sadeh", NULL}

\item{age}{Participant age}

\item{fragmentation}{Calculate fragmentation? (default: TRUE)}

\item{circadian}{Calculate circadian? (default: TRUE)}

\item{exclude_sleep}{Exclude sleep from activity? (default: TRUE)}

\item{parallel}{Use parallel? (default: auto)}

\item{cores}{Number of CPU cores}
}
\value{
Config list for canhrActi.batch()
}
\description{
Helper to create a config object for canhrActi.batch().
Only specify parameters you want to change from defaults.
}
\examples{
\dontrun{
# Create config for CANHR analysis
cfg <- batch.config(wear = "CANHR2025", intensity = "CANHR")
results <- canhrActi.batch("C:/Data", config = cfg)

# Config for sleep analysis
cfg <- batch.config(sleep = "cole_kripke", circadian = TRUE)
}

}
