% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canhrActi-package.R
\docType{package}
\name{canhrActi-package}
\alias{canhrActi-package}
\title{canhrActi: CANHR ActiGraph Physical Activity Analysis}
\description{
Comprehensive accelerometer data analysis suite developed by the Center for Alaska Native Health Research (CANHR). Implements scientifically validated algorithms including the official ActiGraph activity counts algorithm (Neishabouri et al., 2022), Choi (2011) and Troiano (2007) wear time detection, Freedson (1998) intensity cut points, Cole-Kripke and Sadeh sleep scoring, and 12 METs prediction algorithms. Features unique wake period inversion (sleepwaker methodology) for separating sleep from waking sedentary behavior. Includes circadian rhythm analysis (L5, M10, IS, IV, cosinor), sedentary fragmentation metrics, and auto-calibration for raw data. Supports ActiGraph (.agd, .gt3x), GENEActiv (.bin), Axivity (.cwa), and generic CSV files. Produces results compatible with ActiLife software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdazadda/canhrActi}
  \item Report bugs at \url{https://github.com/rdazadda/canhrActi/issues}
}

}
\author{
\strong{Maintainer}: Raymond Dacosta Azadda \email{rdazadda@alaska.edu}

Authors:
\itemize{
  \item Andrew Grogan-Kaylor \email{agrogan@umich.edu}
  \item KyungSook Lee \email{klee98@alaska.edu}
}

}
\keyword{internal}
