% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canhrActi-package.R
\docType{package}
\name{canhrActi-package}
\alias{canhrActi-package}
\title{canhrActi: CANHR ActiGraph Physical Activity Analysis}
\description{
Comprehensive ActiGraph accelerometer data analysis suite developed by the Center for Alaska Native Health Research (CANHR). Implements scientifically validated algorithms including the official ActiGraph activity counts algorithm (Neishabouri et al., 2022), Choi (2011) and Troiano (2007) wear time detection, and Freedson (1998) intensity cut points. Produces results matching ActiLife software. Designed for reproducible physical activity research with comprehensive documentation and user-friendly functions for complete AGD file analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rdazadda/canhrActi}
  \item Report bugs at \url{https://github.com/rdazadda/canhrActi/issues}
}

}
\author{
\strong{Maintainer}: Raymond Dacosta Azadda \email{rdazadda@alaska.edu}

Authors:
\itemize{
  \item Andrew Grogan-Kaylor \email{agrogan@umich.edu}
  \item KyungSook Lee \email{klee98@alaska.edu}
}

}
\keyword{internal}
