% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_devices.R
\name{check.device.support}
\alias{check.device.support}
\title{Check Available Device Support}
\usage{
check.device.support()
}
\value{
Data frame with device types and availability
}
\description{
Lists which accelerometer device formats are supported based on
installed packages.
}
