% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{cpp-interface}
\alias{cpp-interface}
\title{C++ Computational Core Interface}
\description{
R interface functions for the high-performance C++ computational core.
These functions provide 10-50x speedup for computationally intensive
operations on large accelerometer datasets.
}
\keyword{internal}
