% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_sleep.R
\name{.export.actilife.sleep}
\alias{.export.actilife.sleep}
\title{Export Sleep Data to ActiLife Format (Internal)}
\usage{
.export.actilife.sleep(results.list, output_dir)
}
\arguments{
\item{results.list}{List of individual sleep results}

\item{output_dir}{Output directory}
}
\description{
Internal function called by batch_sleep.R to export ActiLife-compatible files.
}
\keyword{internal}
