% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R, R/read_gt3x.R
\name{.is.agd.file}
\alias{.is.agd.file}
\title{Check if AGD File is Raw}
\usage{
.is.agd.file(filepath)

.is.agd.file(filepath)
}
\arguments{
\item{filepath}{Character. Path to file}
}
\value{
Logical. TRUE if raw AGD, FALSE if epoch AGD

Logical. TRUE if file has .agd extension
}
\description{
Quick check to determine if a file is in .agd format based on extension.
}
\keyword{internal}
