% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{evenson}
\alias{evenson}
\title{Apply Evenson (2008) Children Cut Points}
\usage{
evenson(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for children (5-8 years) validated against indirect calorimetry.
Recommended by systematic reviews as most accurate for children.
}
\details{
Evenson (2008) cutpoints for children:
\itemize{
\item Sedentary: 0-100 CPM (< 1.5 METs)
\item Light: 101-2295 CPM (1.5-3.99 METs)
\item Moderate: 2296-4011 CPM (4.0-6.99 METs)
\item Vigorous: >= 4012 CPM (>= 7.0 METs)
}
}
\references{
Evenson KR, et al. (2008). Calibration of two objective measures of physical
activity for children. J Sports Sci, 26(14), 1557-1565.

Trost SG, et al. (2011). Comparison of accelerometer cut points for
predicting activity intensity in youth. Med Sci Sports Exerc, 43(7), 1360-1368.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{chandler}()},
\code{\link{copeland}()},
\code{\link{mattocks}()},
\code{\link{pate_preschool}()},
\code{\link{puyau}()},
\code{\link{romanzini}()}
}
\concept{children cut-points}
