% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_interface.R
\name{fragmentation_cpp}
\alias{fragmentation_cpp}
\title{Sedentary Fragmentation Metrics}
\usage{
fragmentation_cpp(counts, sedentary_threshold = 100L, wear = integer())
}
\arguments{
\item{counts}{Activity counts}

\item{sedentary_threshold}{Sedentary threshold (default: 100)}

\item{wear}{Wear time vector (optional)}
}
\value{
List with ASTP, SATP, W50, Gini, alpha
}
\description{
Sedentary Fragmentation Metrics
}
\references{
Chastin SF, Granat MH (2010). Gait & Posture.
}
