% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{freedson}
\alias{freedson}
\title{Apply Freedson Adult (1998) Cut Points}
\usage{
freedson(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of counts per minute}
}
\value{
Ordered factor with levels: sedentary, light, moderate, vigorous, very_vigorous
}
\description{
Apply Freedson Adult (1998) Cut Points
}
