% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{freedson_vm3}
\alias{freedson_vm3}
\title{Apply Freedson VM3 (2011) Cut Points}
\usage{
freedson_vm3(vm_counts_per_minute)
}
\arguments{
\item{vm_counts_per_minute}{Numeric vector of vector magnitude CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Triaxial cut-points from Freedson's Vector Magnitude validation study.
Alternative to Sasaki VM3 with slightly different thresholds.
}
\details{
Freedson VM3 (2011) cutpoints:
\itemize{
\item Sedentary: 0-99 VM CPM
\item Light: 100-2452 VM CPM
\item Moderate: 2453-6891 VM CPM
\item Vigorous: >= 6892 VM CPM
}
}
\references{
Freedson PS, et al. (2011). Evaluation of artificial neural network algorithms
for predicting METs and activity type from accelerometer data. J Sci Med Sport.
}
\seealso{
Other triaxial cut-points:
\code{\link[=sasaki_vm3]{sasaki_vm3()}}
}
\concept{triaxial cut-points}
