% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{intensity}
\alias{intensity}
\title{Summarize Activity Intensity}
\usage{
intensity(intensity_levels, wear_time = NULL)
}
\arguments{
\item{intensity_levels}{Factor vector from freedson}

\item{wear_time}{Logical vector indicating wear time (optional)}
}
\value{
Data frame with minutes and percentage in each category
}
\description{
Summarize Activity Intensity
}
