% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{matthews}
\alias{matthews}
\title{Apply Matthews (2005) Cut Points}
\usage{
matthews(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Adult cut-points from NHANES III data, validated against doubly labeled water.
Commonly used in epidemiological studies.
}
\details{
Matthews (2005) cutpoints:
\itemize{
\item Sedentary: 0-99 CPM (< 1.5 METs)
\item Light: 100-759 CPM (1.5-2.9 METs)
\item Lifestyle: 760-1951 CPM (3.0-3.9 METs)
\item Moderate: 1952-5724 CPM (4.0-5.9 METs)
\item Vigorous: >= 5725 CPM (>= 6.0 METs)
}
}
\references{
Matthews CE, et al. (2005). Sources of variance in daily physical activity
levels as measured by an accelerometer. Med Sci Sports Exerc, 37(2), 290-298.
}
\seealso{
Other adult cut-points:
\code{\link[=crouter]{crouter()}},
\code{\link[=freedson]{freedson()}},
\code{\link[=santos_lozano]{santos_lozano()}},
\code{\link[=troiano]{troiano()}}
}
\concept{adult cut-points}
