% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{mattocks}
\alias{mattocks}
\title{Apply Mattocks (2007) Children Cut Points}
\usage{
mattocks(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for children (12 years) validated in the ALSPAC cohort study.
Higher moderate threshold than Evenson.
}
\details{
Mattocks (2007) cutpoints:
\itemize{
\item Sedentary: 0-100 CPM
\item Light: 101-3580 CPM
\item Moderate: 3581-6129 CPM
\item Vigorous: >= 6130 CPM
}
}
\references{
Mattocks C, et al. (2007). Calibration of an accelerometer during free-living
activities in children. Int J Pediatr Obes, 2(4), 218-226.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{chandler}()},
\code{\link{copeland}()},
\code{\link{evenson}()},
\code{\link{pate_preschool}()},
\code{\link{puyau}()},
\code{\link{romanzini}()}
}
\concept{children cut-points}
