% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{mvpa}
\alias{mvpa}
\title{Calculate MVPA Minutes}
\usage{
mvpa(intensity_levels, wear_time = NULL, include_vigorous = TRUE)
}
\arguments{
\item{intensity_levels}{Factor or character vector of intensity levels}

\item{wear_time}{Logical vector indicating valid wear time (default: all TRUE)}

\item{include_vigorous}{Logical. Include vigorous activity in MVPA? (default: TRUE)}
}
\value{
Numeric. Total MVPA minutes during wear time
}
\description{
Calculate MVPA Minutes
}
