% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{plot.canhrActi_circadian}
\alias{plot.canhrActi_circadian}
\title{Plot Circadian Rhythm Profile}
\usage{
\method{plot}{canhrActi_circadian}(x, show_cosinor = TRUE, ...)
}
\arguments{
\item{x}{canhrActi_circadian object from circadian.rhythm()}

\item{show_cosinor}{Logical. Overlay cosinor fit? (default: TRUE)}

\item{...}{Additional arguments passed to plot functions}
}
\value{
ggplot object
}
\description{
Creates a 24-hour activity profile plot with L5/M10 windows highlighted.
}
