% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity.R
\name{plot_mvpa_bouts}
\alias{plot_mvpa_bouts}
\title{Plot MVPA Bouts}
\usage{
plot_mvpa_bouts(x, min_bout_length = 10, drop_time_allowance = 2)
}
\arguments{
\item{x}{A \code{canhrActi_analysis} object from \code{canhrActi()}}

\item{min_bout_length}{Integer. Minimum bout duration in minutes (default: 10)}

\item{drop_time_allowance}{Integer. Drop time allowance in minutes (default: 2)}
}
\value{
A ggplot2 object
}
\description{
Visualizes detected MVPA bouts as timeline segments showing when sustained
moderate-to-vigorous activity occurred.
}
\examples{
\dontrun{
results <- canhrActi("participant.agd")
plot_mvpa_bouts(results)
}

}
