% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{puyau}
\alias{puyau}
\title{Apply Puyau (2002) Children Cut Points}
\usage{
puyau(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for children validated using room calorimetry.
Lower thresholds than Evenson, may overestimate MVPA.
}
\details{
Puyau (2002) cutpoints:
\itemize{
\item Sedentary: 0-799 CPM
\item Light: 800-3199 CPM
\item Moderate: 3200-8199 CPM
\item Vigorous: >= 8200 CPM
}
}
\references{
Puyau MR, et al. (2002). Validation and calibration of physical activity
monitors in children. Obes Res, 10(3), 150-157.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{chandler}()},
\code{\link{copeland}()},
\code{\link{evenson}()},
\code{\link{mattocks}()},
\code{\link{pate_preschool}()},
\code{\link{romanzini}()}
}
\concept{children cut-points}
