% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read.agd}
\alias{read.agd}
\title{Read ActiGraph .agd File}
\usage{
read.agd(filepath)
}
\arguments{
\item{filepath}{Path to .agd file}
}
\value{
Data frame with epoch data
}
\description{
Read ActiGraph .agd File
}
