% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_points.R
\name{romanzini}
\alias{romanzini}
\title{Apply Romanzini (2014) Youth Cut Points}
\usage{
romanzini(counts_per_minute)
}
\arguments{
\item{counts_per_minute}{Numeric vector of CPM}
}
\value{
Ordered factor with intensity levels
}
\description{
Cut-points for youth validated in Brazilian adolescents against indirect
calorimetry during various activities.
}
\details{
Romanzini (2014) cutpoints:
\itemize{
\item Sedentary: 0-180 CPM
\item Light: 181-756 CPM
\item Moderate: 757-1111 CPM
\item Vigorous: >= 1112 CPM
}

Note: These thresholds are notably lower than other cut-points.
}
\references{
Romanzini M, et al. (2014). Calibration of accelerometers
in adolescents. Eur J Sport Sci, 14(1), 91-99.
}
\seealso{
Other children cut-points: 
\code{\link{butte_preschool}()},
\code{\link{chandler}()},
\code{\link{copeland}()},
\code{\link{evenson}()},
\code{\link{mattocks}()},
\code{\link{pate_preschool}()},
\code{\link{puyau}()}
}
\concept{children cut-points}
