% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedentary_fragmentation.R
\name{usual.bout.percentiles}
\alias{usual.bout.percentiles}
\title{Calculate Multiple Usual Bout Percentiles}
\usage{
usual.bout.percentiles(bout_durations, percentiles = c(25, 50, 75, 90))
}
\arguments{
\item{bout_durations}{Numeric vector of bout durations}

\item{percentiles}{Percentiles to calculate (default: c(25, 50, 75, 90))}
}
\value{
Named numeric vector with W25, W50, W75, W90
}
\description{
Calculate Multiple Usual Bout Percentiles
}
