% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_counts.R
\name{vm}
\alias{vm}
\title{Vector Magnitude}
\usage{
vm(counts_x, counts_y, counts_z)
}
\arguments{
\item{counts_x}{X-axis counts}

\item{counts_y}{Y-axis counts}

\item{counts_z}{Z-axis counts}
}
\value{
Vector magnitude counts
}
\description{
Vector Magnitude
}
