% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear_time.R
\name{wear.troiano}
\alias{wear.troiano}
\title{Wear Time (Troiano)}
\usage{
wear.troiano(
  counts_per_minute,
  non_wear_window = 60,
  spike_tolerance = 2,
  spike_stoplevel = 100
)
}
\arguments{
\item{counts_per_minute}{Counts per minute}

\item{non_wear_window}{Consecutive zero minutes for non-wear (default: 60)}

\item{spike_tolerance}{Allowed spike minutes in non-wear (default: 2)}

\item{spike_stoplevel}{Maximum CPM for spikes (default: 100)}
}
\value{
Logical vector (TRUE = wear time)
}
\description{
Wear Time (Troiano)
}
